refactor: update school counts to use Long type
Change the data type for school counts from Integer to Long across multiple classes to accommodate larger values.

- Updated method signatures and implementations in ResultWriter, SchoolAnalyzer, and their tests.
- Adjusted logging and CSV reading/writing to reflect the new data type.
- Deprecated DefaultSchoolNameExtractor in favor of improved extraction methods.

+++ src/main/java/com/kakaobank/tools/comment/analyzer/csv/ResultWriter.java
@@ -4,5 +4,5 @@ import java.io.IOException;
 import java.util.Map;
 
 public interface ResultWriter {
-    void writeResults(Map<String, Integer> schoolCounts, String outputFile) throws IOException;
+    void writeResults(Map<String, Long> schoolCounts, String outputFile) throws IOException;
 }
